refactor(gridsat): rename netcdf_projection module to projection
